Fix docker layer repository caching
Add the --pull flag when running in CI so that repository docker layer
caching works.

Fix the k3os version to include the branch when not on a tag instead of
hexadecimal string. Otherwise the contents of the image differs on each
build after a tag which will mess up the layer caching.