feat: use flatcar kernel
Use the flatcar linux kernel instead of building our own

Disable hfsplus on the iso image so that we don't need to rebuild the
kernel with support for it

Disable joilet as it doesn't support symlinks

Don't mount the kernel source anymore as it is not included in the
flatcar image. Not sure why it was included before as there isn't a need
to build anything against it.